Fix lib's API
This streamlines the public API by foregoing the aux type `Encoded`
in favour of just one wrapper type `Enc` which is obtained directly
by encoding into `Matrix3<Q231>` type and encrypting using a `KeyPair`.
This fixed the problems of incorrectly operating on encrypted data
where, with the previous API, it was possible to add/multiply the
encrypted data by `Encoded` data which would yield incorrect results.
Now, `Enc` requires the underlying data to be encoded **and**
encrypted thus avoiding this problem. However, this comes at a
cost where usage of crates such as `nalgebra` is nigh impossible.
`nalgebra` requires a "scalar" type (an entry for `Matrix`) to have
`num_traits::{Zero, One}` defined which is impossible for our `Enc`
type since this would require a statically known `KeyPair` in order
to be able encrypt and decrypt a value of 0 and 1 respectively.

Changes:
* introduce `Enc` type which stores encoded **and** encrypted data
* leave out *only* `Encrypt` and `Decrypt` traits
* remove `linalg.rs` example as it's nontrivial to make it work
  currently
* generate randomized `KeyPair` by default (`Default` trait)
* update docs